Modification de Resource
Ajout de UserResourceController ainsi que modification des routes
