+++ src/Controller/ResourceController.php
@@ -30,68 +30,4 @@ class ResourceController extends AbstractController
             'resource' => $resource,
         ]);
     }
-
-    /**
-     * @Route("/utilisateurs/profil/ressources/nouveau", name="resource_new", methods={"GET","POST"}, priority=5)
-     */
-    public function new(Request $request): Response
-    {
-        $slugify = new Slugify();
-        $resource = new Resource();
-        $form = $this->createForm(ResourceType::class, $resource);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $resource->setUser($this->getUser());
-            $resource->setSlug($slugify->slugify($resource->getName()));
-            $entityManager->persist($resource);
-            $entityManager->flush();
-
-            return $this->redirectToRoute('user_profile');
-        }
-
-        return $this->render('resource/new.html.twig', [
-            'resource' => $resource,
-            'form' => $form->createView(),
-        ]);
-    }
-
-    /**
-     * @Route("/utilisateurs/profil/ressources/{id}", name="resource_edit", methods={"GET","POST"}, priority=3)
-     */
-    public function edit(Request $request, Resource $resource): Response
-    {
-        $form = $this->createForm(ResourceType::class, $resource);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
-
-            return $this->redirectToRoute('user_profile');
-        }
-
-        return $this->render('resource/edit.html.twig', [
-            'resource' => $resource,
-            'form' => $form->createView(),
-        ]);
-    }
-
-    /**
-     * @Route("/utilisateurs/profil/ressources/supprimer/{id}", name="resource_delete", methods={"POST"}, priority=3)
-     */
-    public function delete(Request $request, Resource $resource): Response
-    {
-        if ($this->isCsrfTokenValid('delete'.$resource->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $resource->setLatest(null);
-            foreach ($resource->getPatchNotes() as $patchNote){
-                $entityManager->remove($patchNote);
-            }
-            $entityManager->remove($resource);
-            $entityManager->flush();
-        }
-
-        return $this->redirectToRoute('user_profile');
-    }
 }
