+++ templates/components/_card.html.twig
@@ -20,7 +20,7 @@
         <div class="description">
             {{ resource.description }}
             {% if app.user and app.user.id == resource.user.id  %}
-                <a href="{{ path('resource_edit', { id: resource.id }) }}">Editer</a>
+                <a href="{{ path('user_resource_edit', { id: resource.id }) }}">Editer</a>
                 {{ include('resource/_delete_form.html.twig') }}
             {% endif %}
         </div>
