+++ templates/resource/_delete_form.html.twig
@@ -1,4 +1,4 @@
-<form method="post" action="{{ path('resource_delete', {id: resource.id}) }}" onsubmit="return confirm('Are you sure you want to delete this item?');">
+<form method="post" action="{{ path('user_resource_delete', {id: resource.id}) }}" onsubmit="return confirm('Are you sure you want to delete this item?');">
     <input type="hidden" name="_token" value="{{ csrf_token('delete' ~ resource.id) }}">
     <button class="ui button">Supprimer</button>
 </form>
