+++ templates/user/show.html.twig
@@ -26,7 +26,7 @@
             </div>
         </div>
         {% if self %}
-            <a href="{{ path('resource_new') }}">Ajouter une ressource</a>
+            <a href="{{ path('user_resource_new') }}">Ajouter une ressource</a>
         {% endif %}
     </div>
